requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.1
PyJWT==2.8.0
flask-wtf==1.2.1
wtforms==3.1.2
certifi==2024.2.2
email_validator==2.1.1
gunicorn==21.2.0
requests==2.31.0
